requirements.txt
flask>=3.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
python-engineio>=4.7.0
eventlet>=0.33.0
requests>=2.31.0
gunicorn>=21.0.0
PyJWT>=2.8.0
